Web crawler/requirements.txt
pandas>=2.0.0
requests>=2.31.0
pyyaml>=6.0.1
python-dateutil>=2.9.0.post0
APScheduler>=3.10.4
yfinance>=0.2.40; python_version >= '3.8'
SQLAlchemy>=2.0.30
psycopg2-binary>=2.9.9
